added new Interfaces
HasSpecifications and HasRoles
